feat:command line args all, fix sort prototypes, add sort comparisons

+++ flash_sort.h
@@ -2,6 +2,6 @@
 #define FLASH_SORT_H_
 
 void flashSort(int a[], int n);
-int flashSortComparisions(int a[], int n);
+long long flashSortComparisions(int a[], int n);
 
 #endif
