+++ insertion_sort.h
@@ -2,6 +2,6 @@
 #define INSERTION_SORT_H_
 
 void insertionSort(int a[], int n);
-int insertionSortComparisons(int a[], int n);
+long long insertionSortComparisons(int a[], int n);
 
 #endif
