+++ radix_sort.h
@@ -2,6 +2,6 @@
 #define RADIX_SORT_H_
 
 void radixSort(int a[], int n);
-int radixSortComparisions(int a[], int n);
+long long radixSortComparisions(int a[], int n);
 
 #endif
